requirement.txt
# Deep Learning (Required)
torch>=2.0.0
torchvision>=0.15.0

# Basic Scientific Computing (Required)
numpy>=1.24.0
pillow>=9.5.0
scipy>=1.10.0

# Training Utilities (Required)
tqdm>=4.65.0
pyyaml>=6.0

# Evaluation (Required)
pytorch-fid>=0.3.0

# Optional - for monitoring and visualization
# matplotlib>=3.7.0
# tensorboard>=2.13.0
# tensorboardX>=2.6.0
# opencv-python>=4.8.0

# Optional - for data processing
# pandas>=2.0.0
# sklearn>=1.3.0
# xgboost>=2.0.0